Moves CurrentSystemTimeZone to shared completely
